Remove extra check for module execution context. Move test cases for error returning functions to new module.
